refactor(database): move database configurations to appropriate directory
