+++ src/server.ts
@@ -3,7 +3,7 @@ import express from "express";
 import "express-async-errors";
 import swaggerUi from "swagger-ui-express";
 
-import "./database";
+import "@shared/infra/typeorm";
 import "@shared/container";
 import errorHandler from "@shared/infra/http/middlewares/errorHandler";
